+++ cucumber-java/src/test/java/io/cucumber/java/MethodScannerTest.java
@@ -1,5 +1,6 @@
 package io.cucumber.java;
 
+import io.cucumber.java.en.Given;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -42,6 +43,13 @@ class MethodScannerTest {
     }
 
     @Test
+    void scan_ignores_bridge_methods() throws NoSuchMethodException {
+        Method method = SpecializedReturnType.class.getMethod("test");
+        MethodScanner.scan(SpecializedReturnType.class, backend);
+        assertThat(scanResult, contains(new SimpleEntry<>(method, method.getAnnotations()[0])));
+    }
+
+    @Test
     void scan_ignores_non_instantiable_class() {
         MethodScanner.scan(NonStaticInnerClass.class, backend);
         assertThat(scanResult, empty());
@@ -81,4 +89,17 @@ class MethodScannerTest {
 
     }
 
+    public interface GenericReturnType {
+        Number test();
+
+    }
+
+    public static class SpecializedReturnType implements GenericReturnType {
+
+        @Given("test")
+        public Integer test() {
+            return 1;
+        }
+
+    }
 }