+++ cucumber-picocontainer/src/test/java/io/cucumber/picocontainer/StepDefinitionsWithTransitiveDependencies.java
@@ -1,5 +1,7 @@
 package io.cucumber.picocontainer;
 
+import org.picocontainer.Disposable;
+
 public class StepDefinitionsWithTransitiveDependencies {
 
     final FirstDependency firstDependency;
@@ -8,11 +10,15 @@ public class StepDefinitionsWithTransitiveDependencies {
         this.firstDependency = firstDependency;
     }
 
-    public static class FirstDependency {
+    public static class FirstDependency implements Disposable {
         final SecondDependency secondDependency;
 
         public FirstDependency(SecondDependency secondDependency) {
             this.secondDependency = secondDependency;
+        }
+
+        @Override
+        public void dispose() {
         }
     }